refactor: Merge types/docs from crud/operations.pyi into crud/operations.py
- Merged type hints and docstrings from crudclient/crud/operations.pyi into crudclient/crud/operations.py.

- Deleted crudclient/crud/operations.pyi.
