+++ crudclient/crud/operations.py
@@ -1,10 +1,22 @@
+"""
+Module `operations.py`
+=====================
+
+This module defines the CRUD operations for API resources.
+It provides implementations for list, create, read, update, partial_update, destroy,
+and custom_action operations.
+"""
+
 import json
 import logging
-from typing import Any, Dict
+from typing import TYPE_CHECKING, Any, Dict
 from typing import List as TypingList  # Rename List to avoid conflict
 from typing import Optional, Union
 
 from pydantic import ValidationError as PydanticValidationError
+
+if TYPE_CHECKING:
+    from .base import Crud
 
 from ..exceptions import DataValidationError
 from ..http.utils import redact_json_body
@@ -15,16 +27,52 @@ from .base import T
 logger = logging.getLogger(__name__)
 
 
-def list_operation(self, parent_id: Optional[str] = None, params: Optional[JSONDict] = None) -> Union[JSONList, TypingList[T], ApiResponse]:
+def list_operation(self: "Crud", parent_id: Optional[str] = None, params: Optional[JSONDict] = None) -> Union[JSONList, TypingList[T], ApiResponse]:
+    """
+    Retrieve a list of resources.
+
+    Args:
+        parent_id: Optional ID of the parent resource for nested resources.
+        params: Optional query parameters.
+
+    Returns:
+        Union[JSONList, List[T], ApiResponse]: List of resources.
+
+    Raises:
+        ValueError: If list action is not allowed for this resource.
+        NotFoundError: If the parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "list" not in self.allowed_actions:
         raise ValueError(f"List action not allowed for {self.__class__.__name__}")
 
     endpoint = self._get_endpoint(parent_args=(parent_id,) if parent_id else None)
     response = self.client.get(endpoint, params=params)
-    return self._validate_list_return(response)
+    return self._validate_list_return(response)  # type: ignore[no-any-return]
 
 
-def create_operation(self, data: Union[JSONDict, T], parent_id: Optional[str] = None, params: Optional[JSONDict] = None) -> Union[T, JSONDict]:
+def create_operation(
+    self: "Crud", data: Union[JSONDict, T], parent_id: Optional[str] = None, params: Optional[JSONDict] = None
+) -> Union[T, JSONDict]:
+    """
+    Create a new resource.
+
+    Args:
+        data: The data for the new resource.
+        parent_id: Optional ID of the parent resource for nested resources.
+        params: Optional query parameters.
+
+    Returns:
+        Union[T, JSONDict]: The created resource.
+
+    Raises:
+        ValueError: If create action is not allowed for this resource.
+        DataValidationError: If the input data fails validation.
+        NotFoundError: If the parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "create" not in self.allowed_actions:
         raise ValueError(f"Create action not allowed for {self.__class__.__name__}")
 
@@ -37,7 +85,7 @@ def create_operation(self, data: Union[JSONDict, T], parent_id: Optional[str] = 
         response = self.client.post(endpoint, json=converted_data, params=params)
 
         # Convert the response to a model instance
-        return self._convert_to_model(response)
+        return self._convert_to_model(response)  # type: ignore[no-any-return]
 
     except PydanticValidationError as e:
         # Redact sensitive data before logging or raising
@@ -57,22 +105,60 @@ def create_operation(self, data: Union[JSONDict, T], parent_id: Optional[str] = 
         raise
 
 
-def read_operation(self, resource_id: str, parent_id: Optional[str] = None) -> Union[T, JSONDict]:
+def read_operation(self: "Crud", resource_id: str, parent_id: Optional[str] = None) -> Union[T, JSONDict]:
+    """
+    Retrieve a specific resource.
+
+    Args:
+        resource_id: The ID of the resource to retrieve.
+        parent_id: Optional ID of the parent resource for nested resources.
+
+    Returns:
+        Union[T, JSONDict]: The retrieved resource.
+
+    Raises:
+        ValueError: If read action is not allowed for this resource.
+        NotFoundError: If the resource or parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "read" not in self.allowed_actions:
         raise ValueError(f"Read action not allowed for {self.__class__.__name__}")
 
     endpoint = self._get_endpoint(resource_id, parent_args=(parent_id,) if parent_id else None)
     response = self.client.get(endpoint)
-    return self._convert_to_model(response)
+    return self._convert_to_model(response)  # type: ignore[no-any-return]
 
 
 def update_operation(
-    self,
+    self: "Crud",
     resource_id: Optional[str] = None,
     data: Optional[Union[JSONDict, T]] = None,
     parent_id: Optional[str] = None,
     update_mode: Optional[str] = None,
 ) -> Union[T, JSONDict]:
+    """
+    Update a resource.
+
+    Args:
+        resource_id: The ID of the resource to update. Can be None for non-standard APIs.
+        data: The data to update the resource with.
+        parent_id: Optional parent ID if this is a nested resource.
+        update_mode: The update mode to use. If None, uses the class's _update_mode.
+            Supported modes:
+            - "standard": Standard RESTful update (default)
+            - "no_resource_id": Update without resource ID in URL (e.g., Tripletex company)
+
+    Returns:
+        Union[T, JSONDict]: The updated resource.
+
+    Raises:
+        ValueError: If update action is not allowed for this resource.
+        DataValidationError: If the input data fails validation.
+        NotFoundError: If the resource or parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "update" not in self.allowed_actions:
         raise ValueError(f"Update action not allowed for {self.__class__.__name__}")
 
@@ -103,7 +189,7 @@ def update_operation(
             response = self.client.put(endpoint, json=converted_data)
 
         # Convert the response to a model instance
-        return self._convert_to_model(response)
+        return self._convert_to_model(response)  # type: ignore[no-any-return]
 
     except PydanticValidationError as e:
         # Redact sensitive data before logging or raising
@@ -123,7 +209,25 @@ def update_operation(
         raise
 
 
-def partial_update_operation(self, resource_id: str, data: Union[JSONDict, T], parent_id: Optional[str] = None) -> Union[T, JSONDict]:
+def partial_update_operation(self: "Crud", resource_id: str, data: Union[JSONDict, T], parent_id: Optional[str] = None) -> Union[T, JSONDict]:
+    """
+    Partially update a specific resource.
+
+    Args:
+        resource_id: The ID of the resource to update.
+        data: The partial updated data for the resource.
+        parent_id: Optional ID of the parent resource for nested resources.
+
+    Returns:
+        Union[T, JSONDict]: The updated resource.
+
+    Raises:
+        ValueError: If partial_update action is not allowed for this resource.
+        DataValidationError: If the input data fails validation.
+        NotFoundError: If the resource or parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "partial_update" not in self.allowed_actions:
         raise ValueError(f"Partial update action not allowed for {self.__class__.__name__}")
 
@@ -136,7 +240,7 @@ def partial_update_operation(self, resource_id: str, data: Union[JSONDict, T], p
         response = self.client.patch(endpoint, json=converted_data)
 
         # Convert the response to a model instance
-        return self._convert_to_model(response)
+        return self._convert_to_model(response)  # type: ignore[no-any-return]
 
     except PydanticValidationError as e:
         # Redact sensitive data before logging or raising
@@ -156,7 +260,20 @@ def partial_update_operation(self, resource_id: str, data: Union[JSONDict, T], p
         raise
 
 
-def destroy_operation(self, resource_id: str, parent_id: Optional[str] = None) -> None:
+def destroy_operation(self: "Crud", resource_id: str, parent_id: Optional[str] = None) -> None:
+    """
+    Delete a specific resource.
+
+    Args:
+        resource_id: The ID of the resource to delete.
+        parent_id: Optional ID of the parent resource for nested resources.
+
+    Raises:
+        ValueError: If destroy action is not allowed for this resource.
+        NotFoundError: If the resource or parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     if "destroy" not in self.allowed_actions:
         raise ValueError(f"Destroy action not allowed for {self.__class__.__name__}")
 
@@ -165,11 +282,26 @@ def destroy_operation(self, resource_id: str, parent_id: Optional[str] = None) -
 
 
 def _prepare_request_body_kwargs(
-    self,
+    self: "Crud",
     data: Optional[Union[JSONDict, T]],
     files: Optional[JSONDict],
     content_type: Optional[str],
 ) -> Dict[str, Any]:
+    """
+    Prepare request body keyword arguments based on content type and data.
+
+    Args:
+        data: Optional data to send with the request.
+        files: Optional dictionary of files to upload (for multipart/form-data requests).
+        content_type: Optional content type for the request.
+
+    Returns:
+        Dict[str, Any]: Dictionary of keyword arguments for the request body.
+
+    Raises:
+        TypeError: If the data type is incompatible with the content type.
+        ValueError: If an unsupported content type is provided.
+    """
     request_body_kwargs = {}
 
     # a. Multipart/Form-Data (Files)
@@ -178,8 +310,8 @@ def _prepare_request_body_kwargs(
 
         # If data is also provided (for additional form fields)
         if data is not None:
-            if hasattr(data, "model_dump") and callable(data.model_dump):  # type: ignore[attr-defined]
-                request_body_kwargs["data"] = data.model_dump()  # type: ignore[attr-defined]
+            if hasattr(data, "model_dump") and callable(getattr(data, "model_dump")):
+                request_body_kwargs["data"] = getattr(data, "model_dump")()
             elif isinstance(data, dict):
                 request_body_kwargs["data"] = data
             else:
@@ -188,8 +320,8 @@ def _prepare_request_body_kwargs(
     # b. Application/x-www-form-urlencoded
     elif content_type == "application/x-www-form-urlencoded":
         if data is not None:
-            if hasattr(data, "model_dump") and callable(data.model_dump):  # type: ignore[attr-defined]
-                request_body_kwargs["data"] = data.model_dump()  # type: ignore[attr-defined]
+            if hasattr(data, "model_dump") and callable(getattr(data, "model_dump")):
+                request_body_kwargs["data"] = getattr(data, "model_dump")()
             elif isinstance(data, dict):
                 request_body_kwargs["data"] = data
             else:
@@ -198,15 +330,15 @@ def _prepare_request_body_kwargs(
     # c. Application/json (Default)
     elif files is None and (content_type is None or content_type == "application/json"):
         if data is not None:
-            if hasattr(data, "model_dump") and callable(data.model_dump):  # type: ignore[attr-defined]
-                request_body_kwargs["json"] = data.model_dump()  # type: ignore[attr-defined]
+            if hasattr(data, "model_dump") and callable(getattr(data, "model_dump")):
+                request_body_kwargs["json"] = getattr(data, "model_dump")()
             elif isinstance(data, dict):
                 request_body_kwargs["json"] = data
             else:
                 raise TypeError("For application/json, 'data' must be a dict or a Pydantic model")
         else:
-            # Explicitly set json=None if no data is provided
-            request_body_kwargs["json"] = None
+            # Explicitly set json={} if no data is provided
+            request_body_kwargs["json"] = {}
 
     # d. Unsupported Content-Type with Data
     elif data is not None and content_type is not None:
@@ -216,7 +348,7 @@ def _prepare_request_body_kwargs(
 
 
 def custom_action_operation(
-    self,
+    self: "Crud",
     action: str,
     method: str = "post",
     resource_id: Optional[str] = None,
@@ -226,6 +358,32 @@ def custom_action_operation(
     files: Optional[JSONDict] = None,
     content_type: Optional[str] = None,
 ) -> Union[T, JSONDict, TypingList[JSONDict]]:
+    """
+    Perform a custom action on the resource.
+
+    Args:
+        action: The name of the custom action.
+        method: The HTTP method to use. Defaults to "post".
+        resource_id: Optional resource ID if the action is for a specific resource.
+        parent_id: Optional ID of the parent resource for nested resources.
+        data: Optional data to send with the request.
+        params: Optional query parameters.
+        files: Optional dictionary of files to upload (for multipart/form-data requests).
+        content_type: Optional content type for the request. If not provided, defaults to
+                     "application/json" unless files are provided (which uses multipart/form-data).
+                     Supported values: "application/json", "application/x-www-form-urlencoded".
+
+    Returns:
+        Union[T, JSONDict, List[JSONDict]]: The API response.
+
+    Raises:
+        TypeError: If the parameters are of incorrect types.
+        ValueError: If the HTTP method is invalid, the action is not defined, or an unsupported content_type is provided.
+        DataValidationError: If the input data fails validation.
+        NotFoundError: If the resource or parent resource (if applicable) is not found.
+        NetworkError: If a network-related error occurs during the request.
+        APIError: For other API-related errors (e.g., 4xx/5xx responses).
+    """
     # Runtime type checks for critical parameters
     if not isinstance(action, str):
         raise TypeError(f"Action must be a string, got {type(action).__name__}")
@@ -260,11 +418,11 @@ def custom_action_operation(
         try:
             # Check if the response is a list type
             if hasattr(response, "__iter__") and not isinstance(response, (dict, str, bytes)):
-                return response
+                return response  # type: ignore[no-any-return]
             # Attempt to convert the response. Specific validation/parsing errors
             # (DataValidationError, ResponseParsingError) should be raised from
             # _convert_to_model or its delegates if they occur.
-            return self._convert_to_model(response)
+            return self._convert_to_model(response)  # type: ignore[no-any-return]
         except Exception as e:
             # Log unexpected errors during response conversion in custom actions
             logger.error(f"Unexpected error converting custom action response: {e}", exc_info=True)
